semi-related LCD cleanup

+++ Marlin/src/lcd/extui/mks_ui/wifiSerial_STM32.cpp
@@ -230,7 +230,7 @@ void WifiSerial::_rx_complete_irq(serial_t *obj) {
 
     WRITE(WIFI_IO1_PIN, HIGH);
 
-    rx_buffer_index_t i = (unsigned int)(obj->rx_head + 1) % WIFI_RX_BUF_SIZE;
+    rx_buffer_index_t i = (uint16_t)(obj->rx_head + 1) % WIFI_RX_BUF_SIZE;
 
     // if we should be storing the received character into the location
     // just before the tail (meaning that the head would advance to the
@@ -311,7 +311,7 @@ void WifiSerial::end() {
 }
 
 int WifiSerial::available() {
-  return ((unsigned int)(WIFI_RX_BUF_SIZE + _serial.rx_head - _serial.rx_tail)) % WIFI_RX_BUF_SIZE;
+  return ((uint16_t)(WIFI_RX_BUF_SIZE + _serial.rx_head - _serial.rx_tail)) % WIFI_RX_BUF_SIZE;
 }
 
 //
